make ReadNeedleMeta a thin wrapper
